Restructure my-books directory
